+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `Error::pos()`.
+
 ### Changed
 - Rename `Stream::gen_error_pos` into `Stream::gen_text_pos`.
 - Rename `Stream::gen_error_pos_from` into `Stream::gen_text_pos_from`.